employee_management/requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.11.8
chardet==3.0.4
coverage==5.3
coveralls==2.2.0
Django==3.1.3
django-coverage-plugin==1.8.0
django-filter==2.4.0
djangorestframework==3.12.2
docopt==0.6.2
docutils==0.16
flake8==3.8.4
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.0.1
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.6.0
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1
